refactor(engine): Rename interfaces
Rename Engine to RecordProcessor, because there will be multiple implementations of this interface
Rename EngineContext to RecordProcessorContext to have consistent naming
Rename EngineImpl to Engine
