+++ engine/src/main/java/io/camunda/zeebe/streamprocessor/StreamProcessor.java
@@ -9,8 +9,8 @@ package io.camunda.zeebe.streamprocessor;
 
 import io.camunda.zeebe.db.TransactionContext;
 import io.camunda.zeebe.db.ZeebeDb;
-import io.camunda.zeebe.engine.EngineImpl;
-import io.camunda.zeebe.engine.api.Engine;
+import io.camunda.zeebe.engine.Engine;
+import io.camunda.zeebe.engine.api.RecordProcessor;
 import io.camunda.zeebe.engine.api.StreamProcessorLifecycleAware;
 import io.camunda.zeebe.engine.metrics.StreamProcessorMetrics;
 import io.camunda.zeebe.engine.processing.streamprocessor.LastProcessingPositions;
@@ -118,7 +118,7 @@ public class StreamProcessor extends Actor implements HealthMonitorable, LogReco
   private ActorFuture<LastProcessingPositions> replayCompletedFuture;
   private final Function<LogStreamBatchWriter, TypedStreamWriter> typedStreamWriterFactory;
 
-  private final Engine engine;
+  private final RecordProcessor engine;
 
   protected StreamProcessor(final StreamProcessorBuilder processorBuilder) {
     actorSchedulingService = processorBuilder.getActorSchedulingService();
@@ -140,7 +140,7 @@ public class StreamProcessor extends Actor implements HealthMonitorable, LogReco
     actorName = buildActorName(processorBuilder.getNodeId(), "StreamProcessor", partitionId);
     metrics = new StreamProcessorMetrics(partitionId);
 
-    engine = new EngineImpl(partitionId, zeebeDb, processorBuilder.getEventApplierFactory());
+    engine = new Engine(partitionId, zeebeDb, processorBuilder.getEventApplierFactory());
   }
 
   public static StreamProcessorBuilder builder() {
